feat(ui): launch dark dense beta redesign

+++ backend/app/modules/places/schemas.py
@@ -73,7 +73,7 @@ class PlaceResponse(BaseModel):
     updated_at: datetime = Field(serialization_alias="updatedAt")
     average_rating: float | None = Field(serialization_alias="averageRating")
     rating_count: int = Field(serialization_alias="ratingCount")
-    current_user_rating: int | None = Field(serialization_alias="currentUserRating")
+    current_user_rating: float | None = Field(serialization_alias="currentUserRating")
     current_user_tried: bool = Field(serialization_alias="currentUserTried")
     current_user_list_ids: list[str] = Field(serialization_alias="currentUserListIds")
     current_user_list_names: list[str] = Field(serialization_alias="currentUserListNames")
